SSCS-10493 Address comments and fix PMD issue
